requirements.txt
streamlit==1.24.0
anthropic==0.3.6
python-dotenv==1.0.0